convert genres tab to TypeScript and refactor
