+++ packages/app/app/components/Dashboard/GenresTab/index.tsx
@@ -1,0 +1,59 @@
+import { SmoothImage } from '@nuclear/ui';
+import React from 'react';
+import { Tab } from 'semantic-ui-react';
+import _ from 'lodash';
+
+import genreToIcon from './mapGenres';
+import styles from './styles.scss';
+import { LastfmTopTag } from '@nuclear/core/src/rest/Lastfm.types';
+import { History } from 'history';
+
+const bannedGenres = [
+  'seen live'
+];
+
+type GenresTabProps = {
+  genres: LastfmTopTag[]
+  history: History
+}
+
+const GenresTab: React.FC<GenresTabProps> = ({
+  genres,
+  history
+}) => {
+
+  function onGenreClick(genreName: string) {
+    history.push(`/tag/${genreName}`);
+  }
+
+  return (
+    <Tab.Pane attached={false}>
+      <div className={styles.genre_tab_container}>
+        {
+          _.filter(genres, genre => !_.includes(bannedGenres, genre.name)).map((tag, i) => {
+            return (
+              <div
+                className={styles.genre_container}
+                key={i}
+                onClick={() => onGenreClick(tag.name)}
+              >
+                <div className={styles.genre_overlay}>
+                  <SmoothImage src={`https://picsum.photos/256?random=${i}`} />
+                </div>
+                <div className={styles.genre_name}>
+                  <div
+                    className={styles.svg_icon}
+                    dangerouslySetInnerHTML={{ __html: genreToIcon(tag.name) }}
+                  />
+                  { tag.name }
+                </div>
+              </div>
+            );
+          })
+        }
+      </div>
+    </Tab.Pane>
+  );
+};
+
+export default GenresTab;
